# packages/contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: aos-api
  version: 0.3.0
  description: >
    Stable contract surface per docs/palantier/20_tech/T-API-aos-api稳定契约.md.
    Aligned Wave-0～5 + G-ALIGN (47) — main paths listed; see T-API for full narrative.
servers:
  - url: /v1
paths:
...
  /actions/execute:
    post:
      operationId: actionsExecute
      summary: Write path — Idempotency-Key required; production write only via Draft
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
        - name: X-Allow-Conflicts
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                draftId:
                  type: string
                actionTypeId:
                  type: string
                objectType:
                  type: string
                objectId:
                  type: string
                payload:
                  type: object
                  additionalProperties: true
                proposed:
                  type: object
                  additionalProperties: true
                autoApprove:
                  type: boolean
                  default: false
      responses:
        "200":
          description: ok (proposed draft or production written)
        "400":
          description: missing Idempotency-Key / validation
        "403":
          description: FORBIDDEN markings
        "409":
          description: field conflict
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...